+++ PremiumScraps/CustomEffects/FakeAirhorn.cs
@@ -1,4 +1,6 @@
-﻿using PremiumScraps.Utils;
+﻿using GameNetcodeStuff;
+using PremiumScraps.Utils;
+using System.Collections;
 using Unity.Netcode;
 using UnityEngine;
 
@@ -18,32 +20,53 @@ namespace PremiumScraps.CustomEffects
                 }
                 else  // 30%
                 {
-                    if (Random.Range(1, 11) >= 6)  // 50%
-                    {
-                        AudioServerRpc(6, playerHeldBy.transform.position, 1.5f);  // landmine audio
-                        if (playerHeldBy.IsHost)
-                            StartCoroutine(Effects.DamageHost(playerHeldBy, 100, CauseOfDeath.Burning, (int)Effects.DeathAnimation.Fire));  // death (host)
-                        else
-                            Effects.Damage(playerHeldBy, 100, CauseOfDeath.Burning, (int)Effects.DeathAnimation.Fire);  // death
-                    }
-                    else  // 50%
-                    {
-                        ExplosionServerRpc(playerHeldBy.transform.position);  // explosion
-                    }
+                    var player = playerHeldBy;
+                    StartCoroutine(BadEffect(player, used, buttonDown));
                 }
             }
         }
 
-        [ServerRpc(RequireOwnership = false)]
-        private void BaseItemActivateServerRpc(bool used, bool buttonDown)
+        private IEnumerator BadEffect(PlayerControllerB player, bool used, bool buttonDown = true)
         {
-            BaseItemActivateClientRpc(used, buttonDown);
+            bool explosion = false;
+            if (Random.Range(0, 2) == 0)
+                explosion = true;
+            BaseItemActivateServerRpc(used, buttonDown, true, explosion ? 0.5f : 1.5f);  // airhorn warning audio
+            yield return new WaitForSeconds(0.7f);
+            if (player.isPlayerDead)
+            {
+                RestorePitchServerRpc();
+                yield break;
+            }
+            if (!explosion)  // 50%
+            {
+                AudioServerRpc(6, player.transform.position, 1.5f);  // landmine audio
+                Effects.Damage(player, 100, CauseOfDeath.Burning, (int)Effects.DeathAnimation.Fire);  // death
+            }
+            else  // 50%
+            {
+                ExplosionServerRpc(player.transform.position);  // explosion
+            }
+            yield return new WaitForSeconds(1.2f);
+            RestorePitchServerRpc();
+        }
+
+        [ServerRpc(RequireOwnership = false)]
+        private void BaseItemActivateServerRpc(bool used, bool buttonDown, bool warning = false, float warningPitch = -1)
+        {
+            BaseItemActivateClientRpc(used, buttonDown, warning, warningPitch);
         }
 
         [ClientRpc]
-        private void BaseItemActivateClientRpc(bool used, bool buttonDown)
+        private void BaseItemActivateClientRpc(bool used, bool buttonDown, bool warning, float warningPitch)
         {
-            base.ItemActivate(used, buttonDown);
+            if (!warning)
+                base.ItemActivate(used, buttonDown);
+            else
+            {
+                noiseAudio.pitch = warningPitch;
+                noiseAudio.PlayOneShot(noiseSFX[0], 1f);
+            }
         }
 
         [ServerRpc(RequireOwnership = false)]
@@ -59,6 +82,18 @@ namespace PremiumScraps.CustomEffects
         }
 
         [ServerRpc(RequireOwnership = false)]
+        private void RestorePitchServerRpc()
+        {
+            RestorePitchClientRpc();
+        }
+
+        [ClientRpc]
+        private void RestorePitchClientRpc()
+        {
+            noiseAudio.pitch = noisemakerRandom.Next((int)(minPitch * 100f), (int)(maxPitch * 100f)) / 100f;
+        }
+
+        [ServerRpc(RequireOwnership = false)]
         private void ExplosionServerRpc(Vector3 position)
         {
             ExplosionClientRpc(position);
@@ -67,10 +102,7 @@ namespace PremiumScraps.CustomEffects
         [ClientRpc]
         private void ExplosionClientRpc(Vector3 position)
         {
-            if (IsHost)
-                StartCoroutine(Effects.ExplosionHostDeath(position, 4f, 50, 2));
-            else
-                Effects.Explosion(position, 4f, 50, 2);
+            Effects.Explosion(position, 4.5f, 40, 2);
         }
     }
 }